Module google_sasportal1_alpha1::api

Structs

CustomerDeploymentCreateCall
Creates a new deployment.

CustomerDeploymentDeleteCall
Deletes a deployment.

CustomerDeploymentDeviceCreateCall
Creates a device under a node or customer.

CustomerDeploymentDeviceCreateSignedCall
Creates a signed device under a node or customer.

CustomerDeploymentDeviceListCall
Lists devices under a node or customer.

CustomerDeploymentGetCall
Returns a requested deployment.

CustomerDeploymentListCall
Lists deployments.

CustomerDeploymentMoveCall
Moves a deployment under another node or customer.

CustomerDeploymentPatchCall
Updates an existing deployment.

CustomerDeviceCreateCall
Creates a device under a node or customer.

CustomerDeviceCreateSignedCall
Creates a signed device under a node or customer.

CustomerDeviceDeleteCall
Deletes a device.

CustomerDeviceGetCall
Gets details about a device.

CustomerDeviceListCall
Lists devices under a node or customer.

CustomerDeviceMoveCall
Moves a device under another node or customer.

CustomerDevicePatchCall
Updates a device.

CustomerDeviceSignDeviceCall
Signs a device.

CustomerDeviceUpdateSignedCall
Updates a signed device.

CustomerGetCall
Returns a requested customer.

CustomerListCall
Returns a list of requested customers.

CustomerMethods
A builder providing access to all methods supported on customer resources. It is not used directly, but through the Sasportal hub.

CustomerNodeCreateCall
Creates a new node.

CustomerNodeDeleteCall
Deletes a node.

CustomerNodeDeploymentCreateCall
Creates a new deployment.

CustomerNodeDeploymentListCall
Lists deployments.

CustomerNodeDeviceCreateCall
Creates a device under a node or customer.

CustomerNodeDeviceCreateSignedCall
Creates a signed device under a node or customer.

CustomerNodeDeviceListCall
Lists devices under a node or customer.

CustomerNodeGetCall
Returns a requested node.

CustomerNodeListCall
Lists nodes.

CustomerNodeMoveCall
Moves a node under another node or customer.

CustomerNodeNodeCreateCall
Creates a new node.

CustomerNodeNodeListCall
Lists nodes.

CustomerNodePatchCall
Updates an existing node.

CustomerPatchCall
Updates an existing customer.

DeploymentDeviceDeleteCall
Deletes a device.

DeploymentDeviceGetCall
Gets details about a device.

DeploymentDeviceMoveCall
Moves a device under another node or customer.

DeploymentDevicePatchCall
Updates a device.

DeploymentDeviceSignDeviceCall
Signs a device.

DeploymentDeviceUpdateSignedCall
Updates a signed device.

DeploymentGetCall
Returns a requested deployment.

DeploymentMethods
A builder providing access to all methods supported on deployment resources. It is not used directly, but through the Sasportal hub.

InstallerGenerateSecretCall
Generates a secret to be used with the ValidateInstaller.

InstallerMethods
A builder providing access to all methods supported on installer resources. It is not used directly, but through the Sasportal hub.

InstallerValidateCall
Validates the identity of a Certified Professional Installer (CPI).

NodeDeploymentDeleteCall
Deletes a deployment.

NodeDeploymentDeviceCreateCall
Creates a device under a node or customer.

NodeDeploymentDeviceCreateSignedCall
Creates a signed device under a node or customer.

NodeDeploymentDeviceListCall
Lists devices under a node or customer.

NodeDeploymentGetCall
Returns a requested deployment.

NodeDeploymentListCall
Lists deployments.

NodeDeploymentMoveCall
Moves a deployment under another node or customer.

NodeDeploymentPatchCall
Updates an existing deployment.

NodeDeviceCreateCall
Creates a device under a node or customer.

NodeDeviceCreateSignedCall
Creates a signed device under a node or customer.

NodeDeviceDeleteCall
Deletes a device.

NodeDeviceGetCall
Gets details about a device.

NodeDeviceListCall
Lists devices under a node or customer.

NodeDeviceMoveCall
Moves a device under another node or customer.

NodeDevicePatchCall
Updates a device.

NodeDeviceSignDeviceCall
Signs a device.

NodeDeviceUpdateSignedCall
Updates a signed device.

NodeGetCall
Returns a requested node.

NodeMethods
A builder providing access to all methods supported on node resources. It is not used directly, but through the Sasportal hub.

NodeNodeCreateCall
Creates a new node.

NodeNodeDeleteCall
Deletes a node.

NodeNodeDeploymentCreateCall
Creates a new deployment.

NodeNodeDeploymentListCall
Lists deployments.

NodeNodeDeviceCreateCall
Creates a device under a node or customer.

NodeNodeDeviceCreateSignedCall
Creates a signed device under a node or customer.

NodeNodeDeviceListCall
Lists devices under a node or customer.

NodeNodeGetCall
Returns a requested node.

NodeNodeListCall
Lists nodes.

NodeNodeMoveCall
Moves a node under another node or customer.

NodeNodeNodeCreateCall
Creates a new node.

NodeNodeNodeListCall
Lists nodes.

NodeNodePatchCall
Updates an existing node.

PolicyGetCall
Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.

PolicyMethods
A builder providing access to all methods supported on policy resources. It is not used directly, but through the Sasportal hub.

PolicySetCall
Sets the access control policy on the specified resource. Replaces any existing policy.

PolicyTestCall
Returns permissions that a caller has on the specified resource.

SasPortalAssignment
Associates members with a role.

SasPortalChannelWithScore
The channel with score.

SasPortalCreateSignedDeviceRequest
Request for CreateSignedDevice.

SasPortalCustomer
Entity representing a SAS customer.

SasPortalDeployment
The Deployment.

SasPortalDevice
There is no detailed description.

SasPortalDeviceAirInterface
Information about the device's air interface.

SasPortalDeviceConfig
Information about the device configuration.

SasPortalDeviceGrant
Device grant. It is an authorization provided by the Spectrum Access System to a device to transmit using specified operating parameters after a successful heartbeat by the device.

SasPortalDeviceMetadata
Device data overridable by both SAS Portal and registration requests.

SasPortalDeviceModel
Information about the model of the device.

SasPortalDpaMoveList
An entry in a DPA's move list.

SasPortalEmpty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

SasPortalFrequencyRange
Frequency range from low_frequency to high_frequency.

SasPortalGenerateSecretRequest
Request for GenerateSecret.

SasPortalGenerateSecretResponse
Response for GenerateSecret.

SasPortalGetPolicyRequest
Request message for GetPolicy method.

SasPortalInstallationParams
Information about the device installation parameters.

SasPortalListCustomersResponse
Response for ListCustomers.

SasPortalListDeploymentsResponse
Response for ListDeployments.

SasPortalListDevicesResponse
Response for ListDevices.

SasPortalListNodesResponse
Response for ListNodes.

SasPortalMoveDeploymentRequest
Request for MoveDeployment.

SasPortalMoveDeviceRequest
Request for MoveDevice.

SasPortalMoveNodeRequest
Request for MoveNode.

SasPortalNode
The Node.

SasPortalOperation
This resource represents a long-running operation that is the result of a network API call.

SasPortalPolicy
Defines an access control policy to the resources.

SasPortalSetPolicyRequest
Request message for SetPolicy method.

SasPortalSignDeviceRequest
Request for SignDevice.

SasPortalStatus
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the API Design Guide.

SasPortalTestPermissionsRequest
Request message for TestPermissions method.

SasPortalTestPermissionsResponse
Response message for TestPermissions method.

SasPortalUpdateSignedDeviceRequest
Request for UpdateSignedDevice.

SasPortalValidateInstallerRequest
Request for ValidateInstaller.

SasPortalValidateInstallerResponse
Response for ValidateInstaller.

Sasportal
Central instance to access all Sasportal related resource activities

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.