Struct google_sasportal1_alpha1::api::NodeMethods

pub struct NodeMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on node resources. It is not used directly, but through the Sasportal hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_sasportal1_alpha1 as sasportal1_alpha1;
 
use std::default::Default;
use oauth2;
use sasportal1_alpha1::Sasportal;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = Sasportal::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `deployments_delete(...)`, `deployments_devices_create(...)`, `deployments_devices_create_signed(...)`, `deployments_devices_list(...)`, `deployments_get(...)`, `deployments_list(...)`, `deployments_move(...)`, `deployments_patch(...)`, `devices_create(...)`, `devices_create_signed(...)`, `devices_delete(...)`, `devices_get(...)`, `devices_list(...)`, `devices_move(...)`, `devices_patch(...)`, `devices_sign_device(...)`, `devices_update_signed(...)`, `get(...)`, `nodes_create(...)`, `nodes_delete(...)`, `nodes_deployments_create(...)`, `nodes_deployments_list(...)`, `nodes_devices_create(...)`, `nodes_devices_create_signed(...)`, `nodes_devices_list(...)`, `nodes_get(...)`, `nodes_list(...)`, `nodes_move(...)`, `nodes_nodes_create(...)`, `nodes_nodes_list(...)` and `nodes_patch(...)`
// to build up your call.
let rb = hub.nodes();

Implementations

impl<'a, C> NodeMethods<'a, C>

pub fn deployments_devices_create(
&self,
request: SasPortalDevice,
parent: &str
) -> NodeDeploymentDeviceCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a device under a node or customer.

Arguments

request - No description provided.

parent - Required. The name of the parent resource.

pub fn deployments_devices_create_signed(
&self,
request: SasPortalCreateSignedDeviceRequest,
parent: &str
) -> NodeDeploymentDeviceCreateSignedCall<'a, C>

Create a builder to help you perform the following task:

Creates a signed device under a node or customer.

Arguments

request - No description provided.

parent - Required. The name of the parent resource.

pub fn deployments_devices_list(
&self,
parent: &str
) -> NodeDeploymentDeviceListCall<'a, C>

Create a builder to help you perform the following task:

Lists devices under a node or customer.

Arguments

parent - Required. The name of the parent resource.

pub fn deployments_delete(&self, name: &str) -> NodeDeploymentDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a deployment.

Arguments

name - Required. The name of the deployment.

pub fn deployments_get(&self, name: &str) -> NodeDeploymentGetCall<'a, C>

Create a builder to help you perform the following task:

Returns a requested deployment.

Arguments

name - Required. The name of the deployment.

pub fn deployments_list(&self, parent: &str) -> NodeDeploymentListCall<'a, C>

Create a builder to help you perform the following task:

Lists deployments.

Arguments

parent - Required. The parent resource name, for example, "nodes/1", customer/1/nodes/2.

pub fn deployments_move(
&self,
request: SasPortalMoveDeploymentRequest,
name: &str
) -> NodeDeploymentMoveCall<'a, C>

Create a builder to help you perform the following task:

Moves a deployment under another node or customer.

Arguments

request - No description provided.

name - Required. The name of the deployment to move.

pub fn deployments_patch(
&self,
request: SasPortalDeployment,
name: &str
) -> NodeDeploymentPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates an existing deployment.

Arguments

request - No description provided.

name - Output only. Resource name.

pub fn devices_create(
&self,
request: SasPortalDevice,
parent: &str
) -> NodeDeviceCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a device under a node or customer.

Arguments

request - No description provided.

parent - Required. The name of the parent resource.

pub fn devices_create_signed(
&self,
request: SasPortalCreateSignedDeviceRequest,
parent: &str
) -> NodeDeviceCreateSignedCall<'a, C>

Create a builder to help you perform the following task:

Creates a signed device under a node or customer.

Arguments

request - No description provided.

parent - Required. The name of the parent resource.

pub fn devices_delete(&self, name: &str) -> NodeDeviceDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a device.

Arguments

name - Required. The name of the device.

pub fn devices_get(&self, name: &str) -> NodeDeviceGetCall<'a, C>

Create a builder to help you perform the following task:

Gets details about a device.

Arguments

name - Required. The name of the device.

pub fn devices_list(&self, parent: &str) -> NodeDeviceListCall<'a, C>

Create a builder to help you perform the following task:

Lists devices under a node or customer.

Arguments

parent - Required. The name of the parent resource.

pub fn devices_move(
&self,
request: SasPortalMoveDeviceRequest,
name: &str
) -> NodeDeviceMoveCall<'a, C>

Create a builder to help you perform the following task:

Moves a device under another node or customer.

Arguments

request - No description provided.

name - Required. The name of the device to move.

pub fn devices_patch(
&self,
request: SasPortalDevice,
name: &str
) -> NodeDevicePatchCall<'a, C>

Create a builder to help you perform the following task:

Updates a device.

Arguments

request - No description provided.

name - Output only. The resource path name.

pub fn devices_sign_device(
&self,
request: SasPortalSignDeviceRequest,
name: &str
) -> NodeDeviceSignDeviceCall<'a, C>

Create a builder to help you perform the following task:

Signs a device.

Arguments

request - No description provided.

name - Output only. The resource path name.

pub fn devices_update_signed(
&self,
request: SasPortalUpdateSignedDeviceRequest,
name: &str
) -> NodeDeviceUpdateSignedCall<'a, C>

Create a builder to help you perform the following task:

Updates a signed device.

Arguments

request - No description provided.

name - Required. The name of the device to update.

pub fn nodes_deployments_create(
&self,
request: SasPortalDeployment,
parent: &str
) -> NodeNodeDeploymentCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new deployment.

Arguments

request - No description provided.

parent - Required. The parent resource name where the deployment is to be created.

pub fn nodes_deployments_list(
&self,
parent: &str
) -> NodeNodeDeploymentListCall<'a, C>

Create a builder to help you perform the following task:

Lists deployments.

Arguments

parent - Required. The parent resource name, for example, "nodes/1", customer/1/nodes/2.

pub fn nodes_devices_create(
&self,
request: SasPortalDevice,
parent: &str
) -> NodeNodeDeviceCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a device under a node or customer.

Arguments

request - No description provided.

parent - Required. The name of the parent resource.

pub fn nodes_devices_create_signed(
&self,
request: SasPortalCreateSignedDeviceRequest,
parent: &str
) -> NodeNodeDeviceCreateSignedCall<'a, C>

Create a builder to help you perform the following task:

Creates a signed device under a node or customer.

Arguments

request - No description provided.

parent - Required. The name of the parent resource.

pub fn nodes_devices_list(&self, parent: &str) -> NodeNodeDeviceListCall<'a, C>

Create a builder to help you perform the following task:

Lists devices under a node or customer.

Arguments

parent - Required. The name of the parent resource.

pub fn nodes_nodes_create(
&self,
request: SasPortalNode,
parent: &str
) -> NodeNodeNodeCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new node.

Arguments

request - No description provided.

parent - Required. The parent resource name where the node is to be created.

pub fn nodes_nodes_list(&self, parent: &str) -> NodeNodeNodeListCall<'a, C>

Create a builder to help you perform the following task:

Lists nodes.

Arguments

parent - Required. The parent resource name, for example, "nodes/1".

pub fn nodes_create(
&self,
request: SasPortalNode,
parent: &str
) -> NodeNodeCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new node.

Arguments

request - No description provided.

parent - Required. The parent resource name where the node is to be created.

pub fn nodes_delete(&self, name: &str) -> NodeNodeDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a node.

Arguments

name - Required. The name of the node.

pub fn nodes_get(&self, name: &str) -> NodeNodeGetCall<'a, C>

Create a builder to help you perform the following task:

Returns a requested node.

Arguments

name - Required. The name of the node.

pub fn nodes_list(&self, parent: &str) -> NodeNodeListCall<'a, C>

Create a builder to help you perform the following task:

Lists nodes.

Arguments

parent - Required. The parent resource name, for example, "nodes/1".

pub fn nodes_move(
&self,
request: SasPortalMoveNodeRequest,
name: &str
) -> NodeNodeMoveCall<'a, C>

Create a builder to help you perform the following task:

Moves a node under another node or customer.

Arguments

request - No description provided.

name - Required. The name of the node to move.

pub fn nodes_patch(
&self,
request: SasPortalNode,
name: &str
) -> NodeNodePatchCall<'a, C>

Create a builder to help you perform the following task:

Updates an existing node.

Arguments

request - No description provided.

name - Output only. Resource name.

pub fn get(&self, name: &str) -> NodeGetCall<'a, C>

Create a builder to help you perform the following task:

Returns a requested node.

Arguments

name - Required. The name of the node.