Struct google_sasportal1_alpha1::api::SasPortalInstallationParams

pub struct SasPortalInstallationParams {
    pub antenna_azimuth: Option<i32>,
    pub antenna_beamwidth: Option<i32>,
    pub antenna_downtilt: Option<i32>,
    pub antenna_gain: Option<i32>,
    pub antenna_model: Option<String>,
    pub cpe_cbsd_indication: Option<bool>,
    pub eirp_capability: Option<i32>,
    pub height: Option<f64>,
    pub height_type: Option<String>,
    pub horizontal_accuracy: Option<f64>,
    pub indoor_deployment: Option<bool>,
    pub latitude: Option<f64>,
    pub longitude: Option<f64>,
    pub vertical_accuracy: Option<f64>,
}

Information about the device installation parameters.

This type is not used in any activity, and only used as part of another schema.

Fields
antenna_azimuth: Option<i32>

Boresight direction of the horizontal plane of the antenna in degrees with respect to true north. The value of this parameter is an integer with a value between 0 and 359 inclusive. A value of 0 degrees means true north; a value of 90 degrees means east. This parameter is optional for Category A devices and conditional for Category B devices.

antenna_beamwidth: Option<i32>

3-dB antenna beamwidth of the antenna in the horizontal-plane in degrees. This parameter is an unsigned integer having a value between 0 and 360 (degrees) inclusive; it is optional for Category A devices and conditional for Category B devices.

antenna_downtilt: Option<i32>

Antenna downtilt in degrees and is an integer with a value between -90 and +90 inclusive; a negative value means the antenna is tilted up (above horizontal). This parameter is optional for Category A devices and conditional for Category B devices.

antenna_gain: Option<i32>

Peak antenna gain in dBi. This parameter is an integer with a value between -127 and +128 (dBi) inclusive.

antenna_model: Option<String>

If an external antenna is used, the antenna model is optionally provided in this field. The string has a maximum length of 128 octets.

cpe_cbsd_indication: Option<bool>

If present, this parameter specifies whether the CBSD is a CPE-CBSD or not.

eirp_capability: Option<i32>

This parameter is the maximum device EIRP in units of dBm/10MHz and is an integer with a value between -127 and +47 (dBm/10 MHz) inclusive. If not included, SAS interprets it as maximum allowable EIRP in units of dBm/10MHz for device category.

height: Option<f64>

Device antenna height in meters. When the heightType parameter value is "AGL", the antenna height should be given relative to ground level. When the heightType parameter value is "AMSL", it is given with respect to WGS84 datum.

height_type: Option<String>

Specifies how the height is measured.

horizontal_accuracy: Option<f64>

A positive number in meters to indicate accuracy of the device antenna horizontal location. This optional parameter should only be present if its value is less than the FCC requirement of 50 meters.

indoor_deployment: Option<bool>

Whether the device antenna is indoor or not. true: indoor. false: outdoor.

latitude: Option<f64>

Latitude of the device antenna location in degrees relative to the WGS 84 datum. The allowed range is from -90.000000 to +90.000000. Positive values represent latitudes north of the equator; negative values south of the equator.

longitude: Option<f64>

Longitude of the device antenna location in degrees relative to the WGS 84 datum. The allowed range is from -180.000000 to +180.000000. Positive values represent longitudes east of the prime meridian; negative values west of the prime meridian.

vertical_accuracy: Option<f64>

A positive number in meters to indicate accuracy of the device antenna vertical location. This optional parameter should only be present if its value is less than the FCC requirement of 3 meters.